Fix: 모달창 햄버거 메뉴 state 관리 수정
